# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Chef API - Usuários
  version: 1.0.0

paths:
  /api/v1/usuarios:
    post:
      tags: [Usuario]
      summary: Cadastra um novo usuário
      operationId: criarUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioRequestDTO'
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponseDTO'

    get:
      tags: [Usuario]
      summary: Busca um usuário por nome
      operationId: buscarPorNome
      parameters:
        - in: query
          name: nome
          description: Nome do usuário a ser buscado
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponseDTO'
        '404':
          description: Usuário não encontrado com o nome fornecido.

  /api/v1/usuarios/{id}:
    get:
      tags: [Usuario]
      summary: Busca usuário por ID
      operationId: buscarUsuarioPorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponseDTO'

    delete:
      tags: [Usuario]
      summary: Deleta usuário por ID
      operationId: deletarUsuarioPorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Usuário deletado com sucesso

    put:
      tags: [Usuario]
      summary: Atualiza um usuário existente
      operationId: atualizar
      parameters:
        - in: path
          name: id
          required: true
          description: ID único do usuário (UUID)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        description: Dados do usuário para atualização.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AtualizarUsuarioRequestDTO'
      responses:
        '200':
          description: Usuário atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResponseDTO'
        '404':
          description: Usuário não encontrado.

  /api/v1/usuarios/{id}/senha:
    patch:
      tags: [Usuario]
      summary: Altera a senha de um usuário específico
      operationId: alterarSenha
      parameters:
        - name: id
          in: path
          required: true
          description: ID do usuário (UUID)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        description: Dados para alteração de senha.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlterarSenhaRequestDTO'
      responses:
        '204':
          description: Senha alterada com sucesso
        '400':
          description: Dados inválidos ou senha incorreta
        '404':
          description: Usuário não encontrado

components:
  schemas:
    UsuarioRequestDTO:
      type: object
      properties:
        nome:
          type: string
          example: João Tavares
        email:
          type: string
          example: joao.tavares@example.com
        login:
          type: string
          example: joaotavares
        senha:
          type: string
          example: 123456
        tipo:
          type: string
          description: Tipo de usuário
          enum: [CLIENTE, DONO_RESTAURANTE, ADMIN]
        endereco:
          $ref: '#/components/schemas/EnderecoDTO'
      required: [nome, email, login, senha, tipo, endereco]

    UsuarioResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        email:
          type: string
        login:
          type: string
        tipo:
          type: string
        endereco:
          $ref: '#/components/schemas/EnderecoDTO'
        dataUltimaAlteracao:
          type: string
          format: date-time
      required: [id, nome, email, login, tipo, endereco, dataUltimaAlteracao]

    EnderecoDTO:
      type: object
      properties:
        rua:
          type: string
          example: Rua das Flores
        numero:
          type: string
          example: 123
        cidade:
          type: string
          example: Barbacena
        cep:
          type: string
          example: 36200-000
        uf:
          type: string
          example: MG
      required: [rua, numero, cidade, cep, uf]

    AlterarSenhaRequestDTO:
      type: object
      properties:
        senhaAtual:
          type: string
          example: 123456
        novaSenha:
          type: string
          example: novaSenha123
      required: [senhaAtual, novaSenha]

    AtualizarUsuarioRequestDTO:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
        login:
          type: string
        tipo:
          type: string
          description: Tipo de perfil do usuário
          enum: [CLIENTE, DONO_RESTAURANTE, ADMIN]
        endereco:
          $ref: '#/components/schemas/EnderecoDTO'

tags:
  - name: Usuario
    description: Operações relacionadas aos usuários